Fix handling of async promise errors in the wizard
When in Node.js promise is rejected while nothing is being attached to
the catch block, it's assumed to be an unhandled exception even though
something might attach later. This rewrite changes how error are
propagated in the two async update handlers so that we don't have any
unhandled exceptions.