Resolve rare race condition between node and renderer
In some very rare cases, it could happen that the messages
were send over IPC even though the code to handle them was not
yet loaded in the renderer process. This change makes sure that
we send messages and overall show window only after from both
electron and renderer side window is ready to be shown.